modeling/requirements.txt
pandas
statsmodels
azure-functions